pyproject.toml
[project]
name = "learn"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "allure-pytest>=2.15.3",
    "pytest-playwright>=0.7.2",
]
